Taking capability into account
The management page is now shown top level in admin menu in order to be
accessible to any logged-in user with correct capability.

Page capability is `manage_options` by default but a filter allow to
change that.

If user has `manage_options` capability, API settings are shown. Only
aliases fields otherwise.